requirements.txt
-i https://pypi.org/simple
bcrypt==3.1.7
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
click==7.0
configparser==4.0.2
cryptography==2.8
flask-cors==3.0.8
flask==1.1.1
google-auth==1.10.1
idna==2.8
itsdangerous==1.1.0
jinja2==2.10.3
kubernetes==10.0.1
markupsafe==1.1.1
oauthlib==3.1.0
paramiko==2.7.1
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.19
pymongo==3.10.1
pynacl==1.3.0
python-dateutil==2.8.1
pyyaml==5.3
requests-oauthlib==1.3.0
requests==2.22.0
rsa==4.0
six==1.14.0
urllib3==1.25.7
websocket-client==0.57.0
werkzeug==0.16.0
